backend/requirements.txt
astroid==2.2.5
autopep8==1.6.0
click==8.1.3
colorama==0.4.4
ecdsa==0.13.2
Flask==2.1.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.0
greenlet==1.1.2
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.4.0
MarkupSafe==2.1.1
pyasn1==0.4.8
pycodestyle==2.8.0
python-dotenv==0.20.0
python-jose==3.3.0
rsa==4.8
six==1.12.0
SQLAlchemy==1.4.37
toml==0.10.2
typed-ast==1.4.2
Werkzeug==2.1.2
wrapt==1.11.1
zipp==3.8.0
